+++ persistent-sqlite/ChangeLog.md
@@ -1,3 +1,7 @@
+## Master
+
+Added a `Show` instance for `SqliteConf`.
+
 ## 2.1.2
 
 * Turn on write-ahead log [#363](https://github.com/yesodweb/persistent/issues/363)